+++ example/spec/serial/AsyncOperationRunnerNonExistentLambdaSpec.js
@@ -2,18 +2,16 @@
 
 const uuidv4 = require('uuid/v4');
 
-const { deleteAsyncOperation } = require('@cumulus/api-client/asyncOperations');
+const { createAsyncOperation, deleteAsyncOperation } = require('@cumulus/api-client/asyncOperations');
 const { ecs, s3 } = require('@cumulus/aws-client/services');
 const { randomString } = require('@cumulus/common/test-utils');
 const { getClusterArn, waitForAsyncOperationStatus } = require('@cumulus/integration-tests');
-const { AsyncOperation } = require('@cumulus/api/models');
 const { findAsyncOperationTaskDefinitionForDeployment } = require('../helpers/ecsHelpers');
 const { loadConfig } = require('../helpers/testUtils');
 
 describe('The AsyncOperation task runner running a non-existent lambda function', () => {
   let asyncOperation;
   let asyncOperationId;
-  let asyncOperationModel;
   let asyncOperationsTableName;
   let asyncOperationTaskDefinition;
   let beforeAllFailed = false;
@@ -26,12 +24,6 @@ describe('The AsyncOperation task runner running a non-existent lambda function'
       config = await loadConfig();
 
       asyncOperationsTableName = `${config.stackName}-AsyncOperationsTable`;
-
-      asyncOperationModel = new AsyncOperation({
-        stackName: config.stackName,
-        systemBucket: config.bucket,
-        tableName: asyncOperationsTableName,
-      });
 
       // Find the ARN of the cluster
       cluster = await getClusterArn(config.stackName);
@@ -47,13 +39,15 @@ describe('The AsyncOperation task runner running a non-existent lambda function'
         Body: JSON.stringify([1, 2, 3]),
       }).promise();
 
-      await asyncOperationModel.create({
+      const asyncOperationObject = {
         id: asyncOperationId,
         taskArn: randomString(),
         description: 'Some description',
         operationType: 'ES Index',
         status: 'RUNNING',
-      });
+      };
+
+      await createAsyncOperation({ prefix: config.stackName, asyncOperation: asyncOperationObject });
 
       const runTaskResponse = await ecs().runTask({
         cluster,